pyproject.toml
[project]
name = "liteyukibot-plugin-melobot"
dynamic = ["version"]
description = "Default template for PDM package"
authors = [
    {name = "snowykami", email = "snowykami@outlook.com"},
]
dependencies = [
    "liteyukibot>=6.3.16",
    "melobot[onebot]==3.0.0rc9",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.version]
source = "scm"
tag_filter = "v*"
tag_regex = '^v(?:\D*)?(?P<version>([1-9][0-9]*!)?(0|[1-9][0-9]*)(\.(0|[1-9][0-9]*))*((a|b|c|rc)(0|[1-9][0-9]*))?(\.post(0|[1-9][0-9]*))?(\.dev(0|[1-9][0-9]*))?$)$'


[tool.pdm]
distribution = false
